Fixed flakiness in MapEntryTest#testDefaultShapes
This test is flaky since the method `java.lang.Class.getDeclaredFields` is non-deterministic. `GetDeclaredFilelds`  is used in several methods in the `com.fasterxml.jackson` library (e.g. `objectMapper.writeValueAsString`). 
I changed the test to compare two jsonNode objects instead of two strings, so the specific order of the property would not affect the result. Also, I added checks for both variants of the generated typescript string due to the non-deterministic ordering of the output. 
The other two tests in `MapEntryTest.java` contain the same issues and can be fixed with similar changes in this PR.