feature: Support for sealed classes containing tagged unions
This uses the proposal of https://github.com/FasterXML/jackson-module-kotlin/issues/239 and the general solution of https://github.com/FasterXML/jackson-module-kotlin/pull/240 to import subtypes of a sealed object or interface that are annotated with `@JsonTypeName` and assignable to the container class or interface to create a much less verbose method of generating tagged unions.

Sealed types are supported in kotlin 1.6+ and Java 17+, however, we don't really need the sealed functionality, as we are an offline, one-shot generator, this may work to generate the correct type mappings, but fail to parse on the runtime object mapper side unless the user is using the supported kotlin and java versions.