Update the ModelCompiler.isValidIdentifierPart() to include more valid chars
Since now the class CustomMappingTypeProcessor correctly extracts each generic type argument
(even nested ones), a generic type can have some special chars.
For instance, in a type such as java.util.List<java.util.List<String>>,
the generic type argument is java.util.List<String>,
which has the characters:   .   <   >

Previously, the class was indicating this as an invalid identifier for the type.
But this is a valid one, even for Java or TypeScript.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>